Move models into data folder
- Moves all the sequelize data models into a data folder instead of
  being in 2 different places
