+++ __tests__/routes.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const supertest = require('supertest');
-const { db } = require('../src/models/index');
+const { db } = require('../src/data/index');
 const server = require('../src/server');
 const request = supertest(server.server);
