+++ src/auth/middleware/basicAuth.js
@@ -1,12 +1,12 @@
 'use strict';
 
 const base64 = require('base-64');
-const { users } = require('../../models/index.js');
+const { users } = require('../../data/index.js');
 
 module.exports = async (req, res, next) => {
-  
+
   if (!req.headers.authorization) { return _authError(); }
-  
+
   let basic = req.headers.authorization.split(' ').pop();
   let [user, pass] = base64.decode(basic).split(':');
